Control Scripts/requirements.txt
opencv_contrib_python==4.11.0.86
pyserial==3.5
